Move the run_snia_end2end out of tests
